requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-JWT-Extended==4.6.0
Flask-CORS==4.0.0
Flask-Migrate==4.0.7
PyMySQL==1.1.0
cryptography==41.0.7
redis==5.0.1
gunicorn==21.2.0
gevent==23.9.1
pytz==2023.3
bcrypt==4.1.2
Werkzeug==3.0.1
requests==2.31.0
pandas==2.2.2
mysql-connector-python==9.4.0
numpy>=1.23.2
python-dateutil>=2.8.2